Split workspace package route modules
